+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.23 2017-??-??
+ - 2017-08-10 Fixed bug#[12971](https://bugs.otrs.org/show_bug.cgi?id=12971) - AgentTicketMove - Queue list on AJAX update does not honor ListType setting.
  - 2017-08-09 Fixed bug#[12981](https://bugs.otrs.org/show_bug.cgi?id=12981) - Out of office not shown in dashboard.
  - 2017-08-09 Fixed bug#[13011](https://bugs.otrs.org/show_bug.cgi?id=13011) - Bad expires value in cookie.
  - 2017-08-08 Fixed bug#[11512](https://bugs.otrs.org/show_bug.cgi?id=11512) - ACL Type field restriction on DynamicField change in AgentTicketProcess does not work.